// api/types/PagedModelCourseCategoryResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { CourseCategoryResponse } from "./CourseCategoryResponse.ts";
import type { PageMetadata } from "./PageMetadata.ts";

export type PagedModelCourseCategoryResponse = {
    /**
     * @type array | undefined
    */
    content?: CourseCategoryResponse[];
    /**
     * @type object | undefined
    */
    page?: PageMetadata;
};